Update stopaddiction.py
This program will output the following
Low 
Moderate 
You should stop using social media.
High
Note that this is just a simple program to give you an idea of how you can check your level of 
social media addiction and receive a suggestion to stop using it if your addiction level is high. It 
is ultimately up to you to decide whether or not to follow this suggestion.